Directly set parameters with defaults
Terraform v0.3.7 changed the behavior of GetOK.

"There is a semantic change in GetOk where it will return true if there
is any value in the diff that is non-zero. Before, it would return true
only if there was a value in the diff."

This resulted in the default value of listen_on_any (false) not being
set. The other defaults, enabled (true), were not affected.